+++ basic/player/skills.h
@@ -4,8 +4,8 @@
 void	create();
 mapping	query_skills();
 void	set_skills(mapping sk);
-void	set_skill(string name, int val);
-float	add_skill(string name, float value);
+void	set_skill(string name, float val);
+int	add_skill(string name, float value);
 float	query_skill(string name);
 float	add_skill_val(string name, float val);
 mixed	use_skill(string name, float fac, mixed *data);
